requirements.txt
amqp==2.6.0
anyjson==0.3.3
asgiref==3.2.7
attrs==19.3.0
backcall==0.1.0
bcrypt==3.1.7
billiard==3.6.3.0
cached-property==1.5.1
celery==4.4.4
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
decorator==4.4.2
Django==3.0.6
django-celery-results==1.2.1
django-chartjs==2.1.0
django-extensions==2.2.9
django-nvd3==0.9.7
docker==4.2.1
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.18.2
idna==2.9
ipython==7.15.0
ipython-genutils==0.2.0
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
kombu==4.6.10
MarkupSafe==1.1.1
paramiko==2.7.1
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
psycopg2-binary==2.8.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
PyNaCl==1.4.0
pyrsistent==0.16.0
python-nvd3==0.14.2
python-slugify==1.1.4
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
tablib==2.0.0
texttable==1.6.2
traitlets==4.3.3
Unidecode==1.1.1
urllib3==1.25.9
vine==1.3.0
wcwidth==0.2.3
websocket-client==0.57.0
